Add a privacy manifest  (#8455)
Since we don't include the analytics code in builds for the app store, the only
thing we need to declare is that we call fstat() on files in the app container,
which core uses internally to check the size of Realm files. This is assuming
that if the app logs into Atlas that's something that the app itself has to
declare rather than us, as we don't automatically talk to the server on our own
behalf.

The CocoaPods bundling might not actually work. This is the approach that most
of the SDKs on the mandatory privacy manifest list are taking, but since Apple
isn't actually checking the manifests yet no one knows if they'll accept this
fairly weird way of bundling the xcprivacy file. The less weird way breaks
building pods as non-framework static libraries.